+++ models/utilities/fault_management/include/fault_random_walk.hh
@@ -74,7 +74,7 @@ bool FaultRandomWalk<T>::set_param(const std::string& param_name, double value, 
   } else if (param_name == "max") {
     rand.upper_limit = value;
   } else if (param_name == "seed") {
-    rand.seed = value;
+    rand.seed = static_cast<unsigned long>(value);
   } else {
     return Fault::set_param(param_name, value);
   }
